Aicionei a documentação swagger e melhorei as responses
