+++ src/main/java/ao/multaplus/statusPayment/service/Service.java
@@ -18,23 +18,23 @@ public class Service {
     @Autowired
     private StatusPaymentRepository statusPayment;
 
-    //Listar os Estados
+
     public ResponseEntity<?> listar(){
         List<StatusPayment> all = statusPayment.findAll();
         return new ResponseEntity<>( all, HttpStatus.OK);
     }
 
 
-    //Deletar Os Estads
+
     public ResponseEntity<?> deletar(long id){
         StatusPayment tate = statusPayment.findById(id);
         statusPayment.delete(tate);
-        sms.setmensagem("Estado Alterado Com Sucesso");
+        sms.setmensagem("Estado do Pagamento Deletado Com Sucesso");
         return new ResponseEntity<>(sms, HttpStatus.OK);
     }
 
 
-    //Actualizar os Estados
+
     public ResponseEntity<?> editar(StatusPayment state){
         if(state.getClass().equals("")){
             sms.setmensagem("O estado não pode estar vazio");
@@ -47,7 +47,6 @@ public class Service {
     }
 
 
-    //Salvar os Estados
     public ResponseEntity<?> cadastrar(StatusPayment state){
         if(state.getClass().equals("")){
             sms.setmensagem("O estado não pode estar vazio");
@@ -60,7 +59,6 @@ public class Service {
     }
 
 
-    //Listar Um Estado
     public ResponseEntity<?> buscar(long id){
         StatusPayment status = statusPayment.findById(id);
