+++ src/main/resources/application-dev.properties
@@ -1,12 +1,12 @@
-server.port=8001
+server.port=8080
 spring.jpa.hibernate.ddl-auto=update
 spring.datasource.url=${DB_URL:jdbc:mysql://localhost:3306/db_multaplus?autoReconnect=true&useSSL=false}
-spring.datasource.username=${DB_USERNAME:root}
-spring.datasource.password=${DB_PASSWORD:Admin}
+spring.datasource.username=${DB_USERNAME:admin}
+spring.datasource.password=${DB_PASSWORD:}
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
 #spring.jackson.mapper.property-naming-strategy=com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy
 
-
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
